chore: Reduce sonar code smells. (#2027)
* refactor: fix Sonar issue "Utility classes should not have public constructors"


* refactor: fix Sonar issue "Boxed Boolean should be avoided in boolean expressions"


* refactor: fix Sonar issue "Lambdas should be replaced with method references"


* refactor: fix Sonar issue "Redundant casts should not be used"


* refactor: fix Sonar issue "Test classes should comply with a naming convention"
